# Remove the need for iteration over entities