Resolves #13, #24 

This uses the Hierarchical BitSet mask anding that I have been
experimenting with to allow fast iteration over the key space.

When `joining` components we only visit the entity rows are shared
between the both components.

 Benchmark | Before | After
--- | --- | ---
parallel build | 428,938 ns/iter (+/- 4,011) | 429,104 ns/iter (+/- 5,833)
parallel update | 191,447 ns/iter (+/- 54,436) | 70,627 ns/iter (+/- 18,932)
pos_vel build | 239,626 ns/iter (+/- 2,882) | 276,367 ns/iter (+/- 13,210)
pos_vel build | 257,680 ns/iter (+/- 99,414) | 49,947 ns/iter (+/- 9,426)