@kvark 2016-04-14
Ok, I reviewed the code and made some notes and suggestions. Very impressive work, as always, @csherratt , thank you!

My biggest concern is the complication of the `Storage` type, with the need to not only implement `UnsafeStorage` now, but also to expose the `BitSet`. For some storage types (like the hashmap that we already have), this is a little redundant, since they already know what they have and what not.
Another thing is the general increase in the code complexity... However, the following factors outweigh the concerns:
  - Storages are not implemented that often. Most users will be happy with built-in hashmaps and vectors. A few others can tolerate the increased complexity (when implementing custom storages) since they probably need high performance for their complex cases.
  - The performance win is tremendous

I suppose we should release v0.5 when this code gets cleaned up, merged, and tested. I'm sure some people may still find v0.4 appealing for its minimalistic design, but we'll move forward.