@kvark 2016-04-14
9..8..7..6..5..4..3..2..1.. MERGING
> Nuclear launch detected!